+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/mapper/MinMapperRegistry.java
@@ -38,7 +38,7 @@ public class MinMapperRegistry extends AbstractMapperRegistry {
 
     @Override
     public String provideDescription() {
-        return "统计输入数据中的最小数据并输出，要求数据的全部数据均必须能被解析为double。";
+        return "统计输入数据中的最小数据并输出，要求待处理数据均必须能被解析为double。";
     }
 
     @Override